requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.1
pytest-mock==3.14.1
redis==6.2.0
requests==2.32.4
sseclient-py==1.8.0
urllib3==2.5.0
